Actualización general
Parece que funciona la definción de tempo y compás, el metrónomo, la grabación de taps y su reproducción en loop, viendo lo que sale por el monitor serie.  No he comprobado la cuantización aún.

Voy a bajarme la aplicación necesaria para leer del puerto serie los mensajes MIDI y que reproduzcan en Ableton Live, para comprobar que todo lo descrito funciona efectivamente.

+++ BeatBox.h
@@ -24,7 +24,7 @@ extern Trigger accX;
 extern Trigger accY;
 extern Trigger accZ;
 
-extern float maxTriggerValue;
+extern int maxTriggerValue;
 extern int maxTriggerAxis;
 
 #endif /* BEATBOX_H_ */
